feat(web): reorder navigation, merge Dashboard+Status, add Surveillance route
- Reorder nav: surveillance → cameras → recordings → dashboard → settings
- Create Surveillance.svelte from old Dashboard (camera grid + live thumbnails)
- Rewrite Dashboard.svelte as merged Stats+Status tabbed page (system resources,
  health summary, storage/health/transcoding history tabs)
- Add legacy redirects: #/stats→#/dashboard, #/status→#/dashboard/health
- Update i18n keys with dashboard.tab.* and dashboard.healthSummary
- Add Archive.svelte route
- Remove deprecated Timelapse.svelte (moved into Recordings with view modes)
