+++ web/src/lib/i18n/en.json
@@ -1051,5 +1051,10 @@
   "timelapse.gallery.prevMonth": "Prev",
   "timelapse.gallery.nextMonth": "Next",
   "timelapse.gallery.timeline": "Timeline",
-  "timelapse.gallery.noData": "No data for this period"
+  "timelapse.gallery.noData": "No data for this period",
+  "surveillance.title": "Surveillance",
+  "dashboard.healthSummary": "Camera Health",
+  "dashboard.tab.storage": "Storage Trends",
+  "dashboard.tab.health": "Health History",
+  "dashboard.tab.transcoding": "Transcoding"
 }
